+++ src/components/sections/modalidades/pabellon/styles.scss
@@ -1,0 +1,69 @@
+@import '../../../../scss/settings/vars';
+
+.column-type__title--right {
+
+  &::before {
+    @media ( min-width: $breakpoint-md) {
+      right: 0;
+    }
+  }
+  @media ( min-width: $breakpoint-md) {
+    text-align: right;
+  }
+}
+
+.column-type__text--right {
+  left: 0;
+  text-align: left;
+
+  @media ( min-width: $breakpoint-md) {
+    right: 0;
+    text-align: right;
+  }
+}
+
+.column-type__details--right {
+  @media ( min-width: $breakpoint-md) {
+    text-align: right;
+  }
+}
+
+.glitch {
+  color: $white;
+  display: block;
+  font-size: .8rem;
+  letter-spacing: 3px;
+  text-transform: uppercase;
+  position: relative;
+  &:hover{
+    &:after{
+      animation: glitch 1s infinite linear;
+      background: $black;
+      color: $white;
+      content:"mas detalles +";
+      top: -5;
+      left: -6px;
+      overflow: hidden;
+      position: absolute;
+      right: 0;
+      text-shadow: 1px -1px  1px $turquoise;
+    }
+  
+    &:before{
+      animation: glitch 1s infinite linear;
+      background: $black;
+      color: $white;
+      content:"mas detalles +";
+      top: 0;
+      left: 5px;
+      overflow: hidden;
+      position: absolute;
+      right: 0;
+      text-shadow: -1px 1px  1px $turquoise;
+    }
+  }
+
+  @media ( min-width: $breakpoint-lg) {
+    font-size: $txt-medium;
+  }
+}
