+++ src/components/sections/modalidades/sinestesia/styles.scss
@@ -1,0 +1,100 @@
+@import '../../../../scss/settings/vars';
+
+.column-type {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  margin: 30px 0;
+  
+  @media ( min-width: $breakpoint-md) {
+    flex-direction: row;
+    margin: 70px 30px;
+  }
+}
+
+.column-type__content {
+  max-width: 90%;
+  order: 2;
+  padding: 0 20px; 
+
+  @media ( min-width: $breakpoint-md) {
+    max-width: 60%;
+    order: 1;
+  }
+  @media ( min-width: $breakpoint-lg) {
+    margin: 60px 0;
+    max-width: 45%;
+    padding: 0 40px;
+  }
+}
+
+.column-type__video {
+  max-width: 100%;
+  order: 1;
+  padding: 0 20px;
+
+  @media ( min-width: $breakpoint-md) {
+    max-width: 40%;
+  }
+  @media ( min-width: $breakpoint-lg) {
+    max-width: 55%;
+    padding: 0 40px;
+  }
+}
+
+.content-title {
+  margin: 5px 0 30px;
+  position: relative;
+  width: 100%;
+
+  &::before {
+    background: $white;
+    bottom: -20px;
+    content: "";
+    height: 2px;
+    position: absolute;
+    width: 60px;
+  }
+}
+
+
+.glitch {
+  color: $white;
+  display: block;
+  font-size: .8rem;
+  letter-spacing: 3px;
+  text-transform: uppercase;
+  position: relative;
+
+  &:hover{
+    &:after{
+      animation: glitch 1.5s infinite linear;
+      background: $black;
+      color: $white;
+      content:"mas detalles +";
+      top: -5;
+      left: -6px;
+      overflow: hidden;
+      position: absolute;
+      right: 0;
+      text-shadow: 1px -1px  1px $turquoise;
+    }
+  
+    &:before{
+      animation: glitch 1.5s infinite linear;
+      background: $black;
+      color: $white;
+      content:"mas detalles +";
+      top: 0;
+      left: 5px;
+      overflow: hidden;
+      position: absolute;
+      right: 0;
+      text-shadow: -1px 1px  1px $turquoise;
+    }
+  }
+
+  @media ( min-width: $breakpoint-lg) {
+    font-size: $txt-medium;
+  }
+}
