+++ src/components/sections/modalidades/styles.scss
@@ -1,0 +1,11 @@
+@import '../../../scss/settings/vars.scss';
+
+.modalities {
+  margin: 0 auto;
+  max-width: 1200px;
+  padding: 20px 0 30px;
+
+  @media ( min-width: $breakpoint-lg) {
+    padding: 80px 0;
+  }
+}
